fix(ci): make avg always return DOUBLE
avg was advertised as DOUBLE by the Python relation wrapper but the
kernel emitted the result in the input type (BIGINT for int columns)
using integer division. Reading those bytes as DOUBLE produced
denormalized garbage like 4.33543e-319 instead of 87750.0, failing
test_spark_group_by.py:73.

Fix both implementations:
- aggregate.cpp: kernel signature outputs DOUBLE; avg_merge casts the
  running sum to double before dividing.
- grouped_aggregate.cpp: finalize_state for AVG always returns DOUBLE
  (previously truncated back to the column type).
- Update the C++ logical-plan integration test that asserted BIGINT.

+++ integration/cpp/test/test_collection_logical_plan.cpp
@@ -851,7 +851,7 @@ TEST_CASE("integration::cpp::test_collection::logical_plan") {
             REQUIRE(cur->chunk_data().data[1].type().alias() == "count");
             REQUIRE(cur->chunk_data().data[2].type().type() == types::logical_type::BIGINT);
             REQUIRE(cur->chunk_data().data[2].type().alias() == "sum");
-            REQUIRE(cur->chunk_data().data[3].type().type() == types::logical_type::BIGINT);
+            REQUIRE(cur->chunk_data().data[3].type().type() == types::logical_type::DOUBLE);
             REQUIRE(cur->chunk_data().data[3].type().alias() == "avg");
             REQUIRE(cur->chunk_data().data[4].type().type() == types::logical_type::BIGINT);
             REQUIRE(cur->chunk_data().data[4].type().alias() == "min");
@@ -862,8 +862,8 @@ TEST_CASE("integration::cpp::test_collection::logical_plan") {
                 REQUIRE(cur->chunk_data().value(1, static_cast<size_t>(num)).value<uint64_t>() == 1);
                 REQUIRE(cur->chunk_data().value(2, static_cast<size_t>(num)).value<int64_t>() ==
                         (reversed + 25) * 2 * 10);
-                REQUIRE(cur->chunk_data().value(3, static_cast<size_t>(num)).value<int64_t>() ==
-                        static_cast<int64_t>((reversed + 25) * 2));
+                REQUIRE(cur->chunk_data().value(3, static_cast<size_t>(num)).value<double>() ==
+                        static_cast<double>((reversed + 25) * 2));
                 REQUIRE(cur->chunk_data().value(4, static_cast<size_t>(num)).value<int64_t>() ==
                         (reversed + 25) * 2 * 10);
                 REQUIRE(cur->chunk_data().value(5, static_cast<size_t>(num)).value<int64_t>() ==
